added auth to all product apis

+++ src/main/java/com/lld/ecommercedemo/Controllers/ProductController.java
@@ -3,10 +3,14 @@ package com.lld.ecommercedemo.Controllers;
 import com.lld.ecommercedemo.Exceptions.ProductNotFoundException;
 import com.lld.ecommercedemo.Models.Product;
 import com.lld.ecommercedemo.Services.ProductService;
+import com.lld.ecommercedemo.Utils.AuthUtils;
 import com.lld.ecommercedemo.dtos.ProductDtos.*;
 import com.lld.ecommercedemo.dtos.ResponseStatus;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,15 +19,21 @@ import java.util.List;
 @RequestMapping("/product")
 public class ProductController {
 
-    ProductService productService;
+    private ProductService productService;
+    private AuthUtils authUtils;
 
     @Autowired
-    public ProductController(@Qualifier("selfProductService") ProductService productService) {
+    public ProductController(@Qualifier("selfProductService") ProductService productService, AuthUtils authUtils) {
         this.productService = productService;
+        this.authUtils = authUtils;
     }
 
     @GetMapping("/{id}")
-    public ProductResponseDto getProductById(@PathVariable int id){
+    public ResponseEntity<ProductResponseDto> getProductById(@PathVariable int id, @RequestHeader("Auth") String token){
+        if(!authUtils.validate_token(token)){
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+
         ProductResponseDto responseDto = new ProductResponseDto();
         try{
             Product product = productService.getProductById(id);
@@ -34,11 +44,16 @@ public class ProductController {
             responseDto.setMessage(e.getMessage());
             responseDto.setResponseStatus(ResponseStatus.FAILURE);
         }
-        return responseDto;
+        return new ResponseEntity<>(responseDto, HttpStatus.OK);
     }
 
     @GetMapping("")
-    public ProductsResponseDto getProducts(){
+    public ResponseEntity<ProductsResponseDto> getProducts(@RequestHeader("Auth") String token){
+
+        if(!authUtils.validate_token(token)){
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+
         ProductsResponseDto responseDto = new ProductsResponseDto();
         try{
             List<Product> products = productService.getProducts();
@@ -49,15 +64,19 @@ public class ProductController {
             responseDto.setMessage(e.getMessage());
             responseDto.setResponseStatus(ResponseStatus.FAILURE);
         }
-        return responseDto;
+        return new ResponseEntity<>(responseDto, HttpStatus.CREATED);
     }
 
     @PostMapping("")
-    public CreateProductResponseDto createProduct(@RequestBody CreateProductRequestDto requestDto){
+    public ResponseEntity<CreateProductResponseDto> createProduct(@RequestBody CreateProductRequestDto requestDto, @RequestHeader("Auth") String token){
+        if(!authUtils.validate_token(token)){
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+
         CreateProductResponseDto responseDto = new CreateProductResponseDto();
         try{
             Product product = productService.createProduct(requestDto.getTitle(), requestDto.getPrice(), requestDto.getDescription(),
-                    requestDto.getImage(), requestDto.getCategoryName());
+                    requestDto.getImage(), requestDto.getCategoryName(), requestDto.getQuantity());
             responseDto.setProduct(product);
             responseDto.setResponseStatus(ResponseStatus.SUCCESS);
         }
@@ -65,10 +84,19 @@ public class ProductController {
             responseDto.setMessage(e.getMessage());
             responseDto.setResponseStatus(ResponseStatus.FAILURE);
         }
-        return responseDto;
+        return new ResponseEntity<>(responseDto, HttpStatus.OK);
+    }
+    @PostMapping("/details")
+    public List<Product> getProductsByIds(@RequestBody GetProductsByIdsReqDTO reqDTO){
+        List<Long> productIds = reqDTO.getProductIds();
+        return this.productService.getProductsByIds(productIds);
     }
     @PatchMapping("/updateImage")
-    public ProductResponseDto updateImage(@RequestBody UpdateImageProductReqDto reqDto){
+    public ResponseEntity<ProductResponseDto> updateImage(@RequestBody UpdateImageProductReqDto reqDto, @RequestHeader("Auth") String token){
+        if(!authUtils.validate_token(token)){
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+
         ProductResponseDto responseDto = new ProductResponseDto();
         try{
             Product product = productService.updateImage(reqDto.getId(), reqDto.getImage());
@@ -79,11 +107,15 @@ public class ProductController {
             responseDto.setMessage(e.getMessage());
             responseDto.setResponseStatus(ResponseStatus.FAILURE);
         }
-        return responseDto;
+        return new ResponseEntity<>(responseDto, HttpStatus.OK);
     }
 
     @PatchMapping("/updatePrice")
-    public ProductResponseDto updatePrice(@RequestBody UpdatePriceProductReqDto reqDto){
+    public ResponseEntity<ProductResponseDto> updatePrice(@RequestBody UpdatePriceProductReqDto reqDto, @RequestHeader("Auth") String token){
+        if(!authUtils.validate_token(token)){
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+
         ProductResponseDto responseDto = new ProductResponseDto();
         try{
             Product product = productService.updatePrice(reqDto.getId(), reqDto.getPrice());
@@ -94,11 +126,32 @@ public class ProductController {
             responseDto.setMessage(e.getMessage());
             responseDto.setResponseStatus(ResponseStatus.FAILURE);
         }
-        return responseDto;
+        return new ResponseEntity<>(responseDto, HttpStatus.OK);
+    }
+    @PatchMapping("/updateQuantity")
+    public ResponseEntity<ProductResponseDto> updateQuantity(@RequestBody UpdateQuantityReqDto reqDto, @RequestHeader("Auth") String token){
+        if(!authUtils.validate_token(token)){
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+
+        ProductResponseDto responseDto = new ProductResponseDto();
+        try {
+            Product product = productService.updateQuantity(reqDto.getId(), reqDto.getQuantity());
+            responseDto.setProduct(product);
+            responseDto.setResponseStatus(ResponseStatus.SUCCESS);
+        } catch (ProductNotFoundException e) {
+            responseDto.setMessage(e.getMessage());
+            responseDto.setResponseStatus(ResponseStatus.FAILURE);
+        }
+        return new ResponseEntity<>(responseDto, HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public ProductResponseDto deleteProduct(@PathVariable long id){
+    public ResponseEntity<ProductResponseDto> deleteProduct(@PathVariable long id, @RequestHeader("Auth") String token){
+        if(!authUtils.validate_token(token)){
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+
         ProductResponseDto productResponseDto = new ProductResponseDto();
         try{
             productService.deleteProduct(id);
@@ -107,6 +160,6 @@ public class ProductController {
         catch (Exception e){
             productResponseDto.setResponseStatus(ResponseStatus.FAILURE);
         }
-        return productResponseDto;
+        return new ResponseEntity<>(productResponseDto, HttpStatus.OK);
     }
 }
